Allow script to not modify addresses
Currently the script will always delete A and AAAA records for the domain,
only replacing them if the appropriate parameters are populated.  This
alters the behaviour: if a parameter is missing, the corresponding record
is left as it is; if the parameter is empty, the record is deleted; if
it's a valid address, then the record is replaced.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>