Support ip4addr=auto
This is useful if updating from behind a router that is too stupid to
be able to do dynamic DNS.  This way, the address used will be that
used by the router for NAT.

The implementation is simple.  It will fail if the connection comes
from an IPv6 address.  To ensure the source address is IPv4 use
curl/wget -4.

Signed-off-by: Martin Schwenke <martin@meltin.net>
Signed-off-by: David Gibson <david@gibson.dropbear.id.au>